+++ src/dynamics/geometry.jl
@@ -81,10 +81,10 @@ $(TYPEDFIELDS)
     σ_levels_half::VectorType = default_sigma_coordinates(nlayers)
 
     "σ at full levels, σₖ"
-    σ_levels_full::VectorType = 0.5*(σ_levels_half[2:end] + σ_levels_half[1:end-1])
-
-    "σ level thicknesses, σₖ₊₁ - σₖ"
-    σ_levels_thick::VectorType = σ_levels_half[2:end] - σ_levels_half[1:end-1]
+    σ_levels_full::VectorType = 0.5*(σ_levels_half[2:end] + σ_levels_half[1:end-1])  
+    
+    "σ level thicknesses, σ_k+1/2 - σ_k-1/2"
+    σ_levels_thick::VectorType = σ_levels_half[2:end] - σ_levels_half[1:end-1]      
 
     "log of σ at full levels, include surface (σ=1) as last element"
     ln_σ_levels_full::VectorType = log.(vcat(σ_levels_full, 1))